% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dLALocPareto}
\alias{dLALocPareto}
\title{PDF of the log-affine local Pareto Distribution}
\usage{
dLALocPareto(x, t, alpha_0, gamma = NULL, delta = NULL)
}
\arguments{
\item{x}{Numeric. Value, where the CDF is evaluated.}

\item{t}{Numeric. Threshold of the distribution.}

\item{alpha_0}{Numeric. Local Pareto alpha at t.}

\item{gamma}{Numeric. Parameter \code{gamma}.}

\item{delta}{Numeric. Parameter \code{delta}. Alternative parameter that can be used instead of \code{gamma}.
\code{delta = alpha_0 * gamma * log(2)} is the increase in the local Pareto alpha when doubling the amount.}
}
\value{
The PDF of the log-affine local Pareto distribution with parameters \code{t}, \code{alpha_0} and \code{gamma} at \code{x}.
}
\description{
Calculates the PDF of a log-affine local Pareto distribution
}
\examples{
dLALocPareto(2000, t = 1000, alpha_0 = 2, gamma = 1)
dLALocPareto((1:10)*1000, t = 1000, alpha_0 = 2, delta = 1)

}
