% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dLocalPareto}
\alias{dLocalPareto}
\title{PDF of the local Pareto Distribution}
\usage{
dLocalPareto(x, alpha, t = NULL)
}
\arguments{
\item{alpha}{Function. \code{alpha(x)} is the local Pareto alpha at \code{x}.}

\item{t}{Numeric. Threshold of the local Pareto distribution \code{alpha(x)} is used for \code{x>t}. If NULL then \code{alpha}
is used on \code{[0,infty)}}
}
\value{
PDF of the piecewise Pareto distribution with threshold \code{t} and parameter function \code{alpha} evaluated at \code{x}
}
\description{
Calculates the PDF of a local Pareto Distribution
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
x <- 0:10 * 1000
dLocalPareto(x, alpha, t)

}
