% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLPModel.R
\name{is.PLP_Model}
\alias{is.PLP_Model}
\title{Check if an object is a PLP_Model}
\usage{
is.PLP_Model(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\description{
Checks if the class of an object is 'PLP_Model'
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
PLPM <- PLP_Model(2, alpha, t, dispersion = 2)
PLPM
is.PLP_Model(PLPM)

}
