% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLPModel.R
\name{is.valid.PLP_Model}
\alias{is.valid.PLP_Model}
\title{Check if an object is a valid PPP_Model}
\usage{
is.valid.PLP_Model(x, comment = FALSE)
}
\arguments{
\item{x}{Object to be checked.}

\item{comment}{If FALSE then the function returns a boolean indicating whether \code{x} is a valid PPP_Model. If TRUE then the function returns a comment instead.}
}
\description{
Checks if an object is a PPP_Model object and whether it is valid for the use in functions like \code{Layer_Mean}
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
PLPM <- PLP_Model(2, alpha, t, dispersion = 2)
PLPM
is.valid.PLP_Model(PLPM)
is.valid.PLP_Model(PLPM, comment = TRUE)

}
