% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LALocPareto_Layer_SM}
\alias{LALocPareto_Layer_SM}
\title{Second layer moment of the log-affine local Pareto Distribution}
\usage{
LALocPareto_Layer_SM(
  Cover,
  AttachmentPoint,
  t,
  alpha_0,
  gamma = NULL,
  delta = NULL
)
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer. Use \code{Inf} for unlimited layers.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{t}{Numeric. Threshold of the distribution.}

\item{alpha_0}{Numeric. Local Pareto alpha at t.}

\item{gamma}{Numeric. Parameter \code{gamma}.}

\item{delta}{Numeric. Parameter \code{delta}. Alternative parameter that can be used instead of \code{gamma}.
\code{delta = alpha_0 * gamma * log(2)} is the increase in the local Pareto alpha when doubling the amount.}
}
\value{
The second moment of the loss of the log-affine local Pareto distribution with parameters \code{t}, \code{alpha_0} and \code{gamma} in the layer
\code{Cover} xs \code{AttachmentPoint}
}
\description{
Calculates the second moment of the loss of a log-affine local Pareto
distribution in a reinsurance layer
}
\examples{
LALocPareto_Layer_SM(4000, 1000, t = 1000, alpha_0 = 2, gamma = 1)
LALocPareto_Layer_SM(4000, 1000, t = 1000, alpha_0 = 2, delta = 1)

}
