% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LALocPareto_ML_Estimator}
\alias{LALocPareto_ML_Estimator}
\title{Maximum Likelihood Estimation for the log-affine local Pareto distribution}
\usage{
LALocPareto_ML_Estimator(losses, t, gamma = NULL, delta = NULL, weights = NULL)
}
\arguments{
\item{losses}{Numeric vector. Losses that are used for the ML estimation.}

\item{t}{Numeric. Threshold of the Pareto distribution.}

\item{gamma}{Numeric vector or NULL. If \code{gamma} and \code{delta} are NULL then the parameter \code{gamma} is estimated.}

\item{delta}{Numeric vector or NULL. If \code{gamma} and \code{delta} are NULL then the parameter \code{delta} is estimated.}

\item{weights}{Numeric vector or NULL. Weights for the losses. For instance \code{weights[i] = 2} and \code{weights[j] = 1} for \code{j != i} has the same effect as adding another loss of size \code{loss[i]}.}
}
\value{
Maximum likelihood estimator for the parameters \code{alpha_0} and \code{gamma} of a log-affine local Pareto distribution with threshold \code{t} given the observations \code{losses}
}
\description{
Calculates the maximum likelihood estimator for the parameters alpha_0 and gamma of a log-affine local Pareto distribution
with a known threshold
}
\examples{
losses <- rLALocPareto(100, 1000, 2, 2)
LALocPareto_ML_Estimator(losses, 1000)

}
