% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectiveModelMethods.R
\name{Layer_Mean}
\alias{Layer_Mean}
\title{Expected Loss of a Reinsurance Layer}
\usage{
Layer_Mean(CollectiveModel, Cover = Inf, AttachmentPoint = 0)
}
\arguments{
\item{CollectiveModel}{A collective model object. Currently only \code{PLP_Models} are handled.}

\item{Cover}{Numeric. Cover of the reinsurance layer. Use \code{Inf} for unlimited layers.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}
}
\value{
The expected loss of the layer \code{Cover} xs \code{AttachmentPoint} for the given \code{CollectiveModel}
}
\description{
Calculates the expected loss of a reinsurance layer for a collective model
}
\examples{
t <- 1000
alpha <- function(x) 2 - 1000 / x
PLPM <- PLP_Model(2, alpha, t, dispersion = 2)
Layer_Mean(PLPM, 2000, 2000)

}
