% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LocalPareto_2_PiecewisePareto}
\alias{LocalPareto_2_PiecewisePareto}
\title{Approximate local Pareto distribution by a piecewise Pareto distribution}
\usage{
LocalPareto_2_PiecewisePareto(
  alpha,
  t = NULL,
  rel.tolerance = 1e-04,
  stop.EP = 1e-09,
  stop.t = Inf
)
}
\arguments{
\item{alpha}{Function.}

\item{t}{Numeric. Threshold of the distribution. If \code{t = 0} then \code{t} is derived from \code{alpha}.}

\item{rel.tolerance}{Numeric. Maximum of the relative error of the approximation for the survival function.}

\item{stop.EP}{Numeric. Algorithm stops if the survival function is below \code{stop.EP}.}

\item{stop.t}{Numeric. Algorithm stops if the threshold exceeds \code{stop.t}.}
}
\value{
List with parameters and infos.
}
\description{
Calculates the parameters of an approximating piecewise Pareto distribution
}
\examples{
alpha_0 <- 1.5
t <- 1000
gamma <- 0.5
alpha <- function(x) alpha_0 + alpha_0 * gamma * log(x/t)
LocalPareto_2_PiecewisePareto(t, alpha)

}
