% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LocalPareto_2_PiecewisePareto_OLD}
\alias{LocalPareto_2_PiecewisePareto_OLD}
\title{Approximate local Pareto distribution by a piecewise Pareto distribution (old version)}
\usage{
LocalPareto_2_PiecewisePareto_OLD(alpha, t = NULL)
}
\arguments{
\item{alpha}{Function.}

\item{t}{Numeric. Threshold of the distribution. If \code{t = 0} then \code{t} is derived from \code{alpha}.}
}
\value{
List with parameters and infos.
}
\description{
Calculates the parameters of an approximating piecewise Pareto distribution (old version)
}
\examples{
alpha_0 <- 1.5
t <- 1000
gamma <- 0.5
alpha <- function(x) alpha_0 + alpha_0 * gamma * log(x/t)
LocalPareto_2_PiecewisePareto_OLD(t, alpha)

}
