% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LocalPareto_Layer_Mean}
\alias{LocalPareto_Layer_Mean}
\title{Layer mean of the local Pareto Distribution}
\usage{
LocalPareto_Layer_Mean(Cover, AttachmentPoint, alpha, t = NULL)
}
\arguments{
\item{Cover}{Numeric. Cover of the reinsurance layer. Use \code{Inf} for unlimited layers.}

\item{AttachmentPoint}{Numeric. Attachment point of the reinsurance layer.}

\item{alpha}{Function. \code{alpha(x)} is the local Pareto alpha at \code{x}.}

\item{t}{Numeric. Threshold of the local Pareto distribution \code{alpha(x)} is used for \code{x>t}. If NULL then \code{alpha}
is used on \code{[0,infty)}}
}
\value{
The expected loss of the local Pareto distribution with threshold \code{t} and parameter function \code{alpha} in the layer
\code{Cover} xs \code{AttachmentPoint}
}
\description{
Calculates the expected loss of a local Pareto
distribution in a reinsurance layer
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
Cover <- c(1000, Inf)
AttachmentPoint <- c(1000, 2000)
LocalPareto_Layer_Mean(Cover, AttachmentPoint, alpha, t = 1000)

}
