% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LocalPareto_ML_Estimator_alpha_0}
\alias{LocalPareto_ML_Estimator_alpha_0}
\title{Maximum Likelihood Estimation for the local Pareto distribution}
\usage{
LocalPareto_ML_Estimator_alpha_0(losses, t, beta, weights = NULL)
}
\arguments{
\item{losses}{Numeric vector. Losses that are used for the ML estimation.}

\item{t}{Numeric. Threshold of the local Pareto distribution.}

\item{beta}{Function. Function of local Pareto alphas.}

\item{weights}{Numeric vector or NULL. Weights for the losses. For instance \code{weights[i] = 2} and \code{weights[j] = 1} for \code{j != i} has the same effect as adding another loss of size \code{loss[i]}.}
}
\value{
Maximum likelihood estimator for the parameters \code{alpha_0} of a local Pareto distribution with alpha(x) = alpha_0 * beta(x)
}
\description{
Calculates the maximum likelihood estimator for the alpha_0 if alpha(x) is of the form alpha(x) = alpha_0 * beta(x)
with a known beta
}
\examples{
t <- 1000
beta <- function(x) {2-1000/x}
alpha_0 <- 2
alpha <- function(x) {alpha_0 * beta(x)}
losses <- rLocalPareto(10000, alpha, t)
LocalPareto_ML_Estimator_alpha_0(losses, t, beta)

}
