% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pLocalPareto}
\alias{pLocalPareto}
\title{CDF of the local Pareto Distribution}
\usage{
pLocalPareto(x, alpha, t = NULL)
}
\arguments{
\item{x}{Numeric vector. The CDF is evaluated at \code{x}.}

\item{alpha}{Function. \code{alpha(x)} is the local Pareto alpha at \code{x}.}

\item{t}{Numeric. Threshold of the local Pareto distribution \code{alpha(x)} is used for \code{x>t}. If NULL then \code{alpha}
is used on \code{[0,infty)}}
}
\value{
CDF of the piecewise Pareto distribution with threshold \code{t} and parameter function \code{alpha} evaluated at \code{x}
}
\description{
Calculates the CDF of a local Pareto Distribution
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
x <- 0:10 * 1000
pLocalPareto(x, alpha, t)

}
