% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLPModel.R
\name{PLP_Model}
\alias{PLP_Model}
\title{PLP_Model (Collective Panjer & Local Pareto Model) Object}
\usage{
PLP_Model(
  FQ = NULL,
  alpha,
  t = NULL,
  dispersion = 1,
  Status = 0,
  Comment = "OK"
)
}
\arguments{
\item{FQ}{Numerical. Expected claim count of the collective model.}

\item{alpha}{Function. \code{alpha(x)} is the local Pareto alpha at \code{x}.}

\item{t}{Numerical. Optional thresholds of the local Pareto distribution. If \code{t} is not \code{NULL} then alpha is only used above \code{t}.}

\item{dispersion}{Numerical. Dispersion of the Panjer distribution (i.e. variance to mean ratio).}

\item{Status}{Numerical indicator if a function returns a PPP_Model object: 0 = success, 1 = some information has been ignored, 2 = no solution found}

\item{Comment}{Charakter. An optional comment.}
}
\description{
Constructor function for the PLP_Model object
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
PLPM <- PLP_Model(2, alpha, t, dispersion = 2)
PLPM

}
