% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLPModel.R
\name{print.PLP_Model}
\alias{print.PLP_Model}
\title{Print a PLP_Model Object (Collective Panjer & Local Pareto Model) Object}
\usage{
\method{print}{PLP_Model}(x, ...)
}
\arguments{
\item{x}{PLP_Model object.}

\item{...}{Other arguments, all currently ignored.}
}
\description{
Print method for PLP_Model objects
}
