% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{qLALocPareto}
\alias{qLALocPareto}
\title{Quantile function of the log-affine local Pareto Distribution}
\usage{
qLALocPareto(p, t, alpha_0, gamma = NULL, delta = NULL)
}
\arguments{
\item{p}{Numeric. The function evaluates the quantile function at \code{p}.}

\item{t}{Numeric. Threshold of the distribution.}

\item{alpha_0}{Numeric. Local Pareto alpha at t.}

\item{gamma}{Numeric. Parameter \code{gamma}.}

\item{delta}{Numeric. Parameter \code{delta}. Alternative parameter that can be used instead of \code{gamma}.
\code{delta = alpha_0 * gamma * log(2)} is the increase in the local Pareto alpha when doubling the amount.}
}
\value{
The quantile function of the log-affine local Pareto distribution with parameters \code{t}, \code{alpha_0} and \code{gamma} at \code{p}.
}
\description{
Calculates the quantile function of a log-affine local Pareto distribution
}
\examples{
qLALocPareto(0.5, 1000, 2, gamma = 1)
qLALocPareto(0.5, 1000, 2, delta = 1)

}
