% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{qLocalPareto}
\alias{qLocalPareto}
\title{Quantile function of the local Pareto Distribution}
\usage{
qLocalPareto(p, alpha, t = NULL)
}
\arguments{
\item{p}{Numeric. The function evaluates the quantile function at \code{p}.}

\item{alpha}{Function. \code{alpha(x)} is the local Pareto alpha at \code{x}.}

\item{t}{Numeric. Threshold of the local Pareto distribution \code{alpha(x)} is used for \code{x>t}. If NULL then \code{alpha}
is used on \code{[0,infty)}}
}
\value{
The quantile function of the local Pareto distribution with threshold \code{t} and parameter function \code{alpha} evaluated at \code{p}.
}
\description{
Calculates the quantile function of a local Pareto Distribution
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
p <- 0:10 * 0.1
qLocalPareto(p, alpha, t)

}
