% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rLALocPareto}
\alias{rLALocPareto}
\title{Simulation of the log-affine local Pareto Distribution}
\usage{
rLALocPareto(n, t, alpha_0, gamma = NULL, delta = NULL)
}
\arguments{
\item{n}{Numeric. Number of observations.}

\item{t}{Numeric vector. Thresholds of the log-affine local Pareto distributions}

\item{alpha_0}{Numeric vector. Local Pareto alphas of the log-affine local Pareto distributions at \code{t}.}

\item{gamma}{NULL or Numeric vector. If \code{gamma} is \code{NULL} then \code{delta} is used.}

\item{delta}{NULL or Numeric vector. If \code{gamma} is \code{NULL} then \code{delta} is used.}
}
\value{
A vector of \code{n} samples from the log-affine local Pareto distribution
}
\description{
Generates random deviates of a log-affine local Pareto distribution
}
\examples{
rLALocPareto(100, 1000, 2, 1)
rLALocPareto(100, 1000, 2, delta = 0.1)

}
