% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rLocalPareto}
\alias{rLocalPareto}
\title{Simulation of the local Pareto distribution}
\usage{
rLocalPareto(n, alpha, t = NULL)
}
\arguments{
\item{n}{Numeric. Number of simulations}

\item{alpha}{Function. \code{alpha(x))} is the local Pareto alpha at \code{x}.}

\item{t}{Numeric. Threshold of the local Pareto distribution \code{alpha(x)} is used for \code{x>t}. If NULL then \code{alpha}
is used on \code{[0,infty)}}
}
\value{
A vector of \code{n} samples from the local Pareto distribution with threshold \code{t} and parameter function \code{alpha}
}
\description{
Generates random deviates of a local Pareto distribution
}
\examples{
t <- 1000
alpha <- function(x) {2-1000/x}
rLocalPareto(100, alpha, t)

}
